temporary push

+++ srcs/job_control/job_init.c
@@ -4,14 +4,18 @@
 
 void	init_shell(void)
 {
-	while (tcgetpgrp(STDIN_FILENO) != (g_shell_pgid = getpgrp()))
-		kill(-g_shell_pgid, SIGTTIN);
-	signals_setup();
-	g_shell_pgid = getpid();
-	if (setpgid(g_shell_pgid, g_shell_pgid) < 0)
+	g_shell_is_interactive = isatty(STDIN_FILENO);
+	if (g_shell_is_interactive)
 	{
-		ft_dprintf(2, "Couldn't put the shell in its own process group\n");
-		exit(1);
+		while (tcgetpgrp(STDIN_FILENO) != (g_shell_pgid = getpgrp()))
+			kill(-g_shell_pgid, SIGTTIN);
+		signals_setup();
+		g_shell_pgid = getpid();
+		if (setpgid(g_shell_pgid, g_shell_pgid) < 0)
+		{
+			ft_dprintf(2, "Couldn't put the shell in its own process group\n");
+			exit(1);
+		}
+		tcsetpgrp(STDIN_FILENO, g_shell_pgid);
 	}
-	tcsetpgrp(STDIN_FILENO, g_shell_pgid);
 }
